pyproject.toml
[tool.poetry]
name = "test_hydra"
version = "0.1.0"
description = ""
authors = ["Nicolas <nicolas.chaulet@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
hydra-core = "^0.11.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
